+++ CHANGELOG.md
@@ -2,4 +2,9 @@
 
 ## [0.0.1] - 2019-10-09
 ### Added
+- Support to hide files
+- Support to hide folders
+- Toggle between normal and hidden view
+- Support to get files back to focus
+- Support to get folders back to focus
 - Created project